authorization

+++ users/delete_user.php
@@ -1,6 +1,12 @@
 <?php 
+session_start();
+if($_SESSION['user_role']!="admin")
+{
+    header("Location: ../unauth.php"); 
+    exit();
+}
 $id = $_GET["userid"]; 
 include_once "../connect.php"; 
 $sql = " delete from users where user_id = $id"; 
-mysqli_query( $myconnection, $sql); 
+mysqli_query( $connection, $sql); 
 header("location: users.php");
